Add rollkit + celestia node
