--- docker-compose.yml
@@ -1,13 +1,0 @@
-services:
-  prometheus:
-    image: prom/prometheus
-    ports:
-      - "9090:9090"
-    volumes:
-      - ./prometheus.yml:/etc/prometheus/prometheus.yml
-  grafana:
-    image: grafana/grafana-oss
-    ports:
-      - "3000:3000"
-    volumes:
-      - ./grafana-storage:/var/lib/grafana
