+++ sequencer/.gitignore
@@ -1,0 +1,20 @@
+# Generated by Cargo
+# will have compiled files and executables
+/.DS_Store
+**/target/
+target/
+
+.vscode/
+
+# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
+# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
+Cargo.lock
+
+hello/build
+bin/
+abci.*
+testnet/
+localnet/
+
+# Grafana local storage
+grafana-storage/
