+++ sequencer/prometheus.yml
@@ -1,0 +1,37 @@
+# docker run -p 9090:9090 -v /Users/aminarria/Lambda/starknet_tendermint_sequencer/prometheus.yml:/etc/prometheus/prometheus.yml prom/prometheus
+
+# my global config
+global:
+  scrape_interval: 15s # Set the scrape interval to every 15 seconds. Default is every 1 minute.
+  evaluation_interval: 15s # Evaluate rules every 15 seconds. The default is every 1 minute.
+  # scrape_timeout is set to the global default (10s).
+
+# Alertmanager configuration
+alerting:
+  alertmanagers:
+    - static_configs:
+        - targets:
+          # - alertmanager:9093
+
+# Load rules once and periodically evaluate them according to the global 'evaluation_interval'.
+rule_files:
+  # - "first_rules.yml"
+  # - "second_rules.yml"
+
+# A scrape configuration containing exactly one endpoint to scrape:
+# Here it's Prometheus itself.
+scrape_configs:
+  # The job name is added as a label `job=<job_name>` to any timeseries scraped from this config.
+  - job_name: "prometheus"
+
+    # metrics_path defaults to '/metrics'
+    # scheme defaults to 'http'.
+
+    static_configs:
+      - targets: ["localhost:9090"]
+      - targets: ["host.docker.internal:26660"]
+        labels:
+          groups: 'local-tendermint'
+      - targets: ["5.9.57.44:26660", "5.9.57.45:26660", "5.9.57.89:26660"]
+        labels:
+          groups: 'tendermint'
